created form, included ai controller for form
