+++ server/controllers/questionController.js
@@ -1,0 +1,8 @@
+const db = require("../models");
+
+module.exports = {
+  createQuestions: async (req, res) => {
+    // send to db, include userID and questions.
+    const sendQuestions = await db.Project;
+  },
+};
