+++ server/models/index.js
@@ -1,9 +1,11 @@
 const Alert = require("./alertModel");
 const User = require("./userModel");
 const Project = require("./projectModel");
+const AIQuestions = require("./AIQuestions");
 
 module.exports = {
   Alert,
   User,
-  Project
+  Project,
+  AIQuestions,
 };
